Backend Finished

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async ( req, res ) => {
     try {
@@ -15,9 +16,10 @@ export const watch = async ( req, res ) => {
     // console.log( "Show video", id );
 
     // console.log( id );
-    const video = await Video.findById( id );
+    const video = await Video.findById( id ).populate( "owner" );
     // console.log( video );
-    if ( video === null ) {
+
+    if ( !video ) {
         return res.render( "404", { pageTitle: "Video not found." } );
         //not working above 404 code
     }
@@ -27,21 +29,30 @@ export const watch = async ( req, res ) => {
 
 export const getEdit = async ( req, res ) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const video = await Video.findById( id );
     if ( !video ) {
         return res.status( 404 ).render( "404", { pageTitle: "Video not found." } );
         //not working above 404 code
+    }
+    console.log( typeof video.owner, typeof _id );
+    if ( String( video.owner ) !== String( _id ) ) {
+        return res.status( 403 ).redirect( "/" );
     }
     return res.render( "edit", { pageTitle: `Editing ${ video.title }`, video } );
 };
 
 export const postEdit = async ( req, res ) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists( { _id: id } );
     if ( !video ) {
         return res.status( 404 ).render( "404", { pageTitle: "Video not found." } );
         //not working above 404 code
+    }
+    if ( String( video.owner ) !== String( _id ) ) {
+        return res.status( 403 ).redirect( "/" );
     }
     await video.findByIdAndUpdate( id, {
         title, description, hashtags: hashtags.Video.formathashtags( hashtags )
@@ -58,23 +69,37 @@ export const getUpload = ( req, res ) => {
 };
 
 export const postUpload = async ( req, res ) => {
+    const { user: { _id } } = req.session;
+    const { path: fileUrl } = req.file;
     //add a video to the videos array
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create( {
+        const newVideo = await Video.create( {
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags: hashtags
         } );
+        const user = await User.findById( _id );
+        user.videos.push( newVideo._id );
+        user.save();
         return res.redirect( "/" );
     } catch ( error ) {
         return res.render( "upload", { pageTitle: "Upload Video", errorMessage: error._message } );
     }
-
 };
 
 export const deleteVideo = async ( req, res ) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
+    const video = await Video.findById( id );
+    if ( !video ) {
+        return res.status( 404 ).render( "404", { pageTitle: "Video not found." } );
+    }
+    if ( String( video.owner ) !== String( _id ) ) {
+        return res.status( 403 ).redirect( "/" );
+    }
     await Video.findByIdAndDelete( id );
     console.log( id );
     //delete video
